% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{parseReacExpr}
\alias{parseReacExpr}
\title{Parse Reaction Expression}
\usage{
parseReacExpr(reacStr, pos = 1, quiet = TRUE)
}
\arguments{
\item{reacStr}{reactiong string as given in EXFOR, i.e. of form (TARGET(PROJ,PROCESS)RESIDUAL,...)}

\item{pos}{position in \code{reacStr} from where the parser should start reading.}
}
\value{
Returns a tree realized as recursive list.
        A leaf node is of the structure as explained in \link{parseReacStr}.
        A non-leaf node represents an arithmetic expression and contains the following elements:
        \describe{
          \item{start}{start position of expression in \code{reacStr}}
          \item{end}{end position of expression in \code{reacStr}}
          \item{reac}{the matched part of the reaction string}
          \item{type}{contains the string "expr"}
          \item{ops}{arithmetic operators separating the \code{nodes}}
          \item{nodes}{sub-expressions being part of this expression (list of lists)}
        }
}
\description{
Parse Reaction Expression
}
