% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{reacStrucToStr}
\alias{reacStrucToStr}
\title{Convert Reaction Structure to Reaction String}
\usage{
reacStrucToStr(reacStruc)
}
\arguments{
\item{reacStruc}{A nested list as returned by \link{parseReacExpr}}
}
\value{
Returns a reaction string according to the EXFOR format
}
\description{
Convert Reaction Structure to Reaction String
}
